quraan media player

+++ app/src/main/java/com/example/android/mohammedfadheel/Utilities.java
@@ -4,16 +4,16 @@ public class Utilities {
     public String milliSecondsToTimer(long milliseconds){
         String finalTimerString = "";
         String secondsString = "";
-                int hours = (int) (milliseconds / (1000*60*60));
-                int minutes = (int) (milliseconds % (1000*60*60)) / (1000*60);
-        int seconds = (int) ((milliseconds % (1000*60*60)) % (1000*60) / 1000);
-        if(hours > 0){
+        int hours = (int) (milliseconds / (1000 * 60 * 60));
+        int minutes = (int) (milliseconds % (1000 * 60 * 60)) / (1000 * 60);
+        int seconds = (int) ((milliseconds % (1000 * 60 * 60)) % (1000 * 60) / 1000);
+        if (hours > 0) {
             finalTimerString = hours + ":";
         }
 
         if (seconds < 10) {
             secondsString = "0" + seconds;
-        }else {
+        } else {
                 secondsString = "" + seconds;}
         finalTimerString = finalTimerString + minutes + ":" + secondsString;
         return finalTimerString;
@@ -32,7 +32,7 @@ public class Utilities {
 
     public int progressToTimer(int progress, int totalDuration){
         int currentDuration = 0;
-        totalDuration = (int) (totalDuration / 1000);
+        totalDuration = totalDuration / 1000;
         currentDuration = (int)((((double)progress) / 100) * totalDuration);
 
         return currentDuration * 1000;
